Bruteforcer for the masked individuals added + some mask fixes
Decided to go with a separate short module for each bruteforce
secondary objective. Will help me sleep at night.

Turns out Individual.mask was never regenerated after the values
have been changed in bruteforce interface functions. Fixed that.

	modified:   evolvers/bruteForcePareto.py
	new file:   evolvers/bruteForceParetoMaskConnectionCost.py
	modified:   individuals/integerWeightsSwitchableConnections.py
	modified:   individuals/realVector.py
	modified:   individuals/realWeightsSwitchableConnections.py

+++ individuals/realVector.py
@@ -9,5 +9,5 @@ class Individual(BaseIndividual):
 		self.values = vals[1:]
 
 	def setValuesToZero(self): # for sparse-first search only! Use setValuesToTheFirstSet() for bruteforce applications
-		self.values = np.zeros(self.params['length'], dtype=np.int)
+		self.values = np.zeros(self.params['length'], dtype=np.float)
 		self.renewID()
